% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gradient_descent}
\alias{gradient_descent}
\title{Gradient descent for the \code{LocationScaleRegression} model class}
\usage{
gradient_descent(
  model,
  stepsize = 0.001,
  maxit = 1000,
  abstol = 0.001,
  verbose = FALSE
)
}
\arguments{
\item{model}{A \code{\link{LocationScaleRegression}} object.}

\item{stepsize}{The scaling factor for the gradient.}

\item{maxit}{The maximum number of iterations.}

\item{abstol}{The absolute convergence tolerance. The algorithm stops if the
absolute value of the gradient drops below this value.}

\item{verbose}{Whether to print the progress of the algorithm.}
}
\value{
The updated model, invisibly.
}
\description{
This function optimizes the log-likelihood of the given location-scale
regression model by gradient descent. It has a side effect on the \code{model}
object.
}
\examples{
y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
gradient_descent(model)

}
