% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LocationScaleRegression}
\alias{LocationScaleRegression}
\title{R6 class for location-scale regression models}
\description{
This model class assumes a normally distributed response variable with
one linear predictor for the location (i.e. the mean) and one for the scale
(i.e. the standard deviation). The linear predictors for the location and
the scale are called \eqn{X\beta} and \eqn{Z\gamma} respectively. The scale
uses a log link.
}
\examples{

## ------------------------------------------------
## Method `LocationScaleRegression$new`
## ------------------------------------------------

y <- rnorm(30)
LocationScaleRegression$new(y ~ 1)


## ------------------------------------------------
## Method `LocationScaleRegression$loglik`
## ------------------------------------------------

y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$loglik()


## ------------------------------------------------
## Method `LocationScaleRegression$grad_beta`
## ------------------------------------------------

y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$grad_beta()

## ------------------------------------------------
## Method `LocationScaleRegression$grad_gamma`
## ------------------------------------------------

y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$grad_gamma()
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{beta}}{A numeric vector with the \code{beta} parameters.}

\item{\code{gamma}}{A numeric vector with the \code{gamma} parameters.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocationScaleRegression$new()}}
\item \href{#method-loglik}{\code{LocationScaleRegression$loglik()}}
\item \href{#method-grad_beta}{\code{LocationScaleRegression$grad_beta()}}
\item \href{#method-grad_gamma}{\code{LocationScaleRegression$grad_gamma()}}
\item \href{#method-clone}{\code{LocationScaleRegression$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$new(
  location,
  scale = ~1,
  data = environment(location),
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{location}}{A two-sided formula with the response variable on the
LHS and the predictor for the location (i.e. the mean)
on the RHS.}

\item{\code{scale}}{A one-sided formula with the predictor for the scale
(i.e. the standard deviation) on the RHS.}

\item{\code{data}}{A data frame (or list or environment) in which to evaluate
the \code{location} and \code{scale} formulas.}

\item{\code{...}}{Passed on to \code{\link[stats:model.matrix]{stats::model.matrix()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Create a new \code{LocationScaleRegression} object.
}

\subsection{Returns}{
A \code{LocationScaleRegression} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
LocationScaleRegression$new(y ~ 1)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loglik"></a>}}
\if{latex}{\out{\hypertarget{method-loglik}{}}}
\subsection{Method \code{loglik()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$loglik()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the log-likelihood of a \code{LocationScaleRegression} object
at the current parameter values.
}

\subsection{Returns}{
A single number.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$loglik()

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grad_beta"></a>}}
\if{latex}{\out{\hypertarget{method-grad_beta}{}}}
\subsection{Method \code{grad_beta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$grad_beta()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the gradient of the log-likelihood of a
\code{LocationScaleRegression} object with respect to \eqn{\beta}
at the current parameter values.
}

\subsection{Returns}{
A numeric vector.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$grad_beta()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-grad_gamma"></a>}}
\if{latex}{\out{\hypertarget{method-grad_gamma}{}}}
\subsection{Method \code{grad_gamma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$grad_gamma()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the gradient of the log-likelihood of a
\code{LocationScaleRegression} object with respect to \eqn{\gamma}
at the current parameter values.
}

\subsection{Returns}{
A numeric vector.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
model <- LocationScaleRegression$new(y ~ 1)
model$grad_gamma()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegression$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
