% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asp20boost.R
\docType{package}
\name{asp20boost}
\alias{asp20boost}
\title{Gradient Boosting for Location-Scale Regression Models}
\usage{
example_model <- LocationScaleRegressionBoost(formula, ...)
gradient_boost(example_model, ...)
}
\description{
The package provides an R6-class \code{LocationScaleRegressionBoost} for model
specification, and a function \code{gradient_boost} for boosting the specified model object.
}
\details{
This package resulted from a series of student projects. Thus, it is part of a
bundle of packages all starting with \code{asp20...}. The overall goal of these packages
is to estimate location-scale regression models, while each project/package employs
different methods. All packages, including \code{asp20boost} build on the central
package \code{asp20model}. Here, an R6-class is specified, which is inherited and expanded
in the \code{asp20boost}-package.
}
\author{
Johannes Strauß, Levin Wiebelt, Sebastiàn Aristizabal
}
