% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{gradient_boost}
\alias{gradient_boost}
\title{Gradient bost for the \code{LocationScaleRegressionBoost} model class}
\usage{
gradient_boost(
  model,
  stepsize = 0.001,
  maxit = 1000,
  abstol = 0.001,
  verbose = TRUE
)
}
\arguments{
\item{model}{A \code{\link{LocationScaleRegressionBoost}} object.}

\item{stepsize}{The scaling factor for the gradient.}

\item{maxit}{The maximum number of iterations.}

\item{abstol}{The absolute convergence tolerance. The algorithm stops if the
absolute value of the gradient drops below this value.}

\item{verbose}{Whether to print the progress of the algorithm.}
}
\value{
The updated model, invisibly.
}
\description{
This function optimizes the log-likelihood of the given location-scale
regression model by gradient descent. It has a side effect on the \code{model}
object.
}
\examples{
y <- rnorm(30)
model <- LocationScaleRegressionBoost$new(y ~ 1)
gradient_boost(model)

}
