% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LocationScaleRegressionBoost}
\alias{LocationScaleRegressionBoost}
\title{asp20boost}
\description{
This model class builds on the class "LocationScaleRegression" of the package
asp20model. It provides measures used by the gradient_boost algorithm to enable
parameter estimation and variable selection.
The model class assumes a normal response, linear precidtors for
location (expectation), as well as linear predictors for scale (variance).
The scale uses the log as a link function.
The linear predictors for the location and
the scale are called \eqn{X\beta} and \eqn{Z\gamma} respectively.
}
\section{Super class}{
\code{\link[asp20model:LocationScaleRegression]{asp20model::LocationScaleRegression}} -> \code{LocationScaleRegression}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{beta}}{A numeric vector with the \code{beta} parameters.}

\item{\code{gamma}}{A numeric vector with the \code{gamma} parameters.}

\item{\code{fitted_location}}{A numeric vector with the fitted values
for the location.}

\item{\code{fitted_scale}}{A numeric vector with the fitted values
for the scale.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{beta}}{A numeric vector with the \code{beta} parameters.}

\item{\code{gamma}}{A numeric vector with the \code{gamma} parameters.}

\item{\code{fitted_location}}{A numeric vector with the fitted values
for the location.}

\item{\code{fitted_scale}}{A numeric vector with the fitted values
for the scale.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-bestFittingVariableGamma}{\code{LocationScaleRegressionBoost$bestFittingVariableGamma()}}
\item \href{#method-bestFittingVariableBeta}{\code{LocationScaleRegressionBoost$bestFittingVariableBeta()}}
\item \href{#method-clone}{\code{LocationScaleRegressionBoost$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_beta">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_beta}{\code{asp20model::LocationScaleRegression$grad_beta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_gamma">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_gamma}{\code{asp20model::LocationScaleRegression$grad_gamma()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="initialize">}\href{../../asp20model/html/LocationScaleRegression.html#method-initialize}{\code{asp20model::LocationScaleRegression$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="loglik">}\href{../../asp20model/html/LocationScaleRegression.html#method-loglik}{\code{asp20model::LocationScaleRegression$loglik()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="resid">}\href{../../asp20model/html/LocationScaleRegression.html#method-resid}{\code{asp20model::LocationScaleRegression$resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bestFittingVariableGamma"></a>}}
\if{latex}{\out{\hypertarget{method-bestFittingVariableGamma}{}}}
\subsection{Method \code{bestFittingVariableGamma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$bestFittingVariableGamma()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the loss - squared error of the partial deviance for each covariate of Gamma


Returns the loss - squared error of the partial deviance for each covariate of Beta


Determins the best variable for componentwise boosting
}

\subsection{Returns}{
Vector of sum of squared errors for each covariate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-bestFittingVariableBeta"></a>}}
\if{latex}{\out{\hypertarget{method-bestFittingVariableBeta}{}}}
\subsection{Method \code{bestFittingVariableBeta()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$bestFittingVariableBeta()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Determins the best variable for componentwise boosting
}

\subsection{Returns}{
New Beta Vector with updated values on the best fitted covariate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
