% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LocationScaleRegressionBoost}
\alias{LocationScaleRegressionBoost}
\title{R6 class for location-scale regression models}
\description{
This model class assumes a normally distributed response variable with
one linear predictor for the location (i.e. the mean) and one for the scale
(i.e. the standard deviation). The linear predictors for the location and
the scale are called \eqn{X\beta} and \eqn{Z\gamma} respectively. The scale
uses a log link.
}
\section{Super class}{
\code{\link[asp20model:LocationScaleRegression]{asp20model::LocationScaleRegression}} -> \code{LocationScaleRegression}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{beta}}{A numeric vector with the \code{beta} parameters.}

\item{\code{gamma}}{A numeric vector with the \code{gamma} parameters.}

\item{\code{fitted_location}}{A numeric vector with the fitted values
for the location.}

\item{\code{fitted_scale}}{A numeric vector with the fitted values
for the scale.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{beta}}{A numeric vector with the \code{beta} parameters.}

\item{\code{gamma}}{A numeric vector with the \code{gamma} parameters.}

\item{\code{fitted_location}}{A numeric vector with the fitted values
for the location.}

\item{\code{fitted_scale}}{A numeric vector with the fitted values
for the scale.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocationScaleRegressionBoost$new()}}
\item \href{#method-clone}{\code{LocationScaleRegressionBoost$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_beta">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_beta}{\code{asp20model::LocationScaleRegression$grad_beta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_gamma">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_gamma}{\code{asp20model::LocationScaleRegression$grad_gamma()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="loglik">}\href{../../asp20model/html/LocationScaleRegression.html#method-loglik}{\code{asp20model::LocationScaleRegression$loglik()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="resid">}\href{../../asp20model/html/LocationScaleRegression.html#method-resid}{\code{asp20model::LocationScaleRegression$resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$new(
  location,
  scale = ~1,
  data = environment(location),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
