% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{LocationScaleRegressionBoost}
\alias{LocationScaleRegressionBoost}
\title{R6 Class for Specifying Location-Scale Regression Models}
\usage{
example_model <- LocationScaleRegressionBoost$new(formula, ...)
}
\description{
Model objects created by this class may be estimated by the \code{gradient_boost} function.
This model class builds upon and thus inherits the basic structure
from the \code{LocationScaleRegression} model class of the \code{asp20model} package.
Consequently, it shares similar syntax and structure. (Please  see the help
page of that class \emph{Note: It would be nice to just link the vignette})
}
\details{
It assumes a normally distributed response variable with
one linear predictor eta_mu for the location (i.e. the mean) and one linear predictor
eta_sigma for scale (i.e. the standard deviation). Scale is modeled by a log-link.
}
\examples{

## ------------------------------------------------
## Method `LocationScaleRegressionBoost$gradients_loglik_mu`
## ------------------------------------------------

y <- rnorm(30)
model <- LocationScaleRegressionBoost$new(y ~ 1)
model$gradients_loglik_mu()
}
\section{Super class}{
\code{\link[asp20model:LocationScaleRegression]{asp20model::LocationScaleRegression}} -> \code{LocationScaleRegressionBoost}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{gradients_loglik_mu}}{Caclulates the unit-wise gradients of the loss-functions,
which is the negative loglikelihood.}

\item{\code{gradients_loglik_sigma}}{Caclulates the unit-wise gradients of the loss-functions,
which is the negative loglikelihood.}

\item{\code{update_parameters_conventional}}{ccc}

\item{\code{update_parameters_compwise}}{ddd}

\item{\code{par_log}}{eee}

\item{\code{plot}}{fff}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{gradients_loglik_mu}}{Caclulates the unit-wise gradients of the loss-functions,
which is the negative loglikelihood.}

\item{\code{gradients_loglik_sigma}}{Caclulates the unit-wise gradients of the loss-functions,
which is the negative loglikelihood.}

\item{\code{update_parameters_conventional}}{ccc}

\item{\code{update_parameters_compwise}}{ddd}

\item{\code{plot}}{fff}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LocationScaleRegressionBoost$new()}}
\item \href{#method-eta_mu}{\code{LocationScaleRegressionBoost$eta_mu()}}
\item \href{#method-eta_sigma}{\code{LocationScaleRegressionBoost$eta_sigma()}}
\item \href{#method-gradients_loglik_mu}{\code{LocationScaleRegressionBoost$gradients_loglik_mu()}}
\item \href{#method-gradients_loglik_sigma}{\code{LocationScaleRegressionBoost$gradients_loglik_sigma()}}
\item \href{#method-update_parameters_conventional}{\code{LocationScaleRegressionBoost$update_parameters_conventional()}}
\item \href{#method-update_parameters_compwise}{\code{LocationScaleRegressionBoost$update_parameters_compwise()}}
\item \href{#method-plot}{\code{LocationScaleRegressionBoost$plot()}}
\item \href{#method-clone}{\code{LocationScaleRegressionBoost$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_beta">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_beta}{\code{asp20model::LocationScaleRegression$grad_beta()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="grad_gamma">}\href{../../asp20model/html/LocationScaleRegression.html#method-grad_gamma}{\code{asp20model::LocationScaleRegression$grad_gamma()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="loglik">}\href{../../asp20model/html/LocationScaleRegression.html#method-loglik}{\code{asp20model::LocationScaleRegression$loglik()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="asp20model" data-topic="LocationScaleRegression" data-id="resid">}\href{../../asp20model/html/LocationScaleRegression.html#method-resid}{\code{asp20model::LocationScaleRegression$resid()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$new(
  location,
  scale = ~1,
  data = environment(location),
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eta_mu"></a>}}
\if{latex}{\out{\hypertarget{method-eta_mu}{}}}
\subsection{Method \code{eta_mu()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$eta_mu()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-eta_sigma"></a>}}
\if{latex}{\out{\hypertarget{method-eta_sigma}{}}}
\subsection{Method \code{eta_sigma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$eta_sigma()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gradients_loglik_mu"></a>}}
\if{latex}{\out{\hypertarget{method-gradients_loglik_mu}{}}}
\subsection{Method \code{gradients_loglik_mu()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$gradients_loglik_mu()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Returns the unit-wise gradients of the loss function, which is the negative loglikelihood.
}

\subsection{Returns}{
a numeric vector nx1
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{y <- rnorm(30)
model <- LocationScaleRegressionBoost$new(y ~ 1)
model$gradients_loglik_mu()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-gradients_loglik_sigma"></a>}}
\if{latex}{\out{\hypertarget{method-gradients_loglik_sigma}{}}}
\subsection{Method \code{gradients_loglik_sigma()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$gradients_loglik_sigma()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_parameters_conventional"></a>}}
\if{latex}{\out{\hypertarget{method-update_parameters_conventional}{}}}
\subsection{Method \code{update_parameters_conventional()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$update_parameters_conventional()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a numeric vector nx1
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-update_parameters_compwise"></a>}}
\if{latex}{\out{\hypertarget{method-update_parameters_compwise}{}}}
\subsection{Method \code{update_parameters_compwise()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$update_parameters_compwise()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a numeric vector nx1
enables plotting
enables plotting
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$plot()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LocationScaleRegressionBoost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
